requirements.txt
Flask==2.3.3
Flask-MySQL==1.5.2
Werkzeug==2.3.7
stripe==7.4.0
reportlab==4.0.4
itsdangerous==2.1.2
bcrypt==4.0.1
python-dotenv==1.0.0
Pillow==10.0.1
email-validator==2.0.0
Flask-OAuthlib==0.9.6
google-auth==2.22.0
google-auth-oauthlib==1.0.0
requests-oauthlib==1.3.1